# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CryptoVault Exchange & Shop API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /packages/my:
    get:
      operationId: listMyPackages
      tags: [packages]
      summary: List current user's purchased packages
      responses:
        "200":
          description: User package list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserPackage"
        "401":
          description: Unauthorized
  /packages/purchase:
    post:
      operationId: purchasePackage
      tags: [packages]
      summary: Purchase a package
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PackagePurchaseRequest"
      responses:
        "201":
          description: Package purchased
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPackage"
        "400":
          description: Insufficient balance or invalid package
        "401":
          description: Unauthorized
  /income:
    get:
      operationId: listIncome
      tags: [income]
      summary: List income history for current user
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
        - name: type
          in: query
          schema: { type: string, enum: [roi, referral, royalty] }
      responses:
        "200":
          description: Income entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/IncomeEntry"
        "401":
          description: Unauthorized
  /income/summary:
    get:
      operationId: getIncomeSummary
      tags: [income]
      summary: Get income summary for current user
      responses:
        "200":
          description: Income summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncomeSummary"
        "401":
          description: Unauthorized
  /admin/packages:
    get:
      operationId: adminListPackages
      tags: [packages]
      summary: Admin list all packages
      responses:
        "200":
          description: Package list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Package"
    post:
      operationId: adminCreatePackage
      tags: [packages]
      summary: Admin create a package
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminPackageInput"
      responses:
        "201":
          description: Package created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
  /admin/packages/{id}:
    put:
      operationId: adminUpdatePackage
      tags: [packages]
      summary: Admin update package config
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminPackageInput"
      responses:
        "200":
          description: Updated package
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
  /admin/referral-levels:
    get:
      operationId: adminGetReferralLevels
      tags: [packages]
      summary: Admin get all 10 referral level configs
      responses:
        "200":
          description: Referral levels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReferralLevel"
    put:
      operationId: adminUpdateReferralLevels
      tags: [packages]
      summary: Admin bulk update referral levels
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReferralLevelsBulkUpdate"
      responses:
        "200":
          description: Updated levels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReferralLevel"
...
components:
  schemas:
...
    Package:
      type: object
      required: [id, name, priceUsdt, roiPercent, roiDays, isActive, createdAt]
      properties:
        id: { type: string }
        name: { type: string }
        priceUsdt: { type: string }
        roiPercent: { type: string }
        roiDays: { type: integer }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    UserPackage:
      type: object
      required: [id, userId, packageId, purchasedAt, daysCredited, totalRoiDays, roiPercent, principalUsdt, totalRoiCredited, isActive]
      properties:
        id: { type: string }
        userId: { type: string }
        packageId: { type: string }
        purchasedAt: { type: string, format: date-time }
        daysCredited: { type: integer }
        totalRoiDays: { type: integer }
        roiPercent: { type: string }
        principalUsdt: { type: string }
        totalRoiCredited: { type: string }
        isActive: { type: boolean }
        lastCreditedAt: { type: ["string", "null"], format: date-time }
    PackagePurchaseRequest:
      type: object
      required: [packageId]
      properties:
        packageId: { type: string }
    AdminPackageInput:
      type: object
      properties:
        name: { type: string }
        priceUsdt: { type: string }
        roiPercent: { type: string }
        roiDays: { type: integer }
        isActive: { type: boolean }
    ReferralLevel:
      type: object
      required: [level, percent, requiredDirects]
      properties:
        level: { type: integer }
        percent: { type: string }
        requiredDirects: { type: integer }
    ReferralLevelsBulkUpdate:
      type: object
      required: [levels]
      properties:
        levels:
          type: array
          items:
            type: object
            required: [level, percent, requiredDirects]
            properties:
              level: { type: integer }
              percent: { type: string }
              requiredDirects: { type: integer }
    IncomeEntry:
      type: object
      required: [id, userId, type, amount, createdAt]
      properties:
        id: { type: string }
        userId: { type: string }
        type:
          type: string
          enum: [roi, referral, royalty]
        amount: { type: string }
        sourceId: { type: ["string", "null"] }
        fromUserId: { type: ["string", "null"] }
        note: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    IncomeSummary:
      type: object
      required: [roi, referral, royalty, activePackages, packages]
      properties:
        roi: { type: string }
        referral: { type: string }
        royalty: { type: string }
        activePackages: { type: integer }
        packages:
          type: array
          items:
            $ref: "#/components/schemas/UserPackage"
...