# lib/api-zod/src/generated/types/incomeEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CryptoVault Exchange & Shop API specification
 * OpenAPI spec version: 0.1.0
 */
import type { IncomeEntryType } from './incomeEntryType';

export interface IncomeEntry {
  id: string;
  userId: string;
  type: IncomeEntryType;
  amount: string;
  /** @nullable */
  sourceId?: string | null;
  /** @nullable */
  fromUserId?: string | null;
  /** @nullable */
  note?: string | null;
  createdAt: Date;
}