# lib/api-zod/src/generated/types/incomeSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CryptoVault Exchange & Shop API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserPackage } from './userPackage';

export interface IncomeSummary {
  roi: string;
  referral: string;
  royalty: string;
  activePackages: number;
  packages: UserPackage[];
}