# lib/api-zod/src/generated/types/package.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CryptoVault Exchange & Shop API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Package {
  id: string;
  name: string;
  priceUsdt: string;
  roiPercent: string;
  roiDays: number;
  isActive: boolean;
  createdAt: Date;
  updatedAt?: Date;
}