# lib/api-zod/src/generated/types/referralLevel.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CryptoVault Exchange & Shop API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReferralLevel {
  level: number;
  percent: string;
  requiredDirects: number;
}