# lib/api-zod/src/generated/types/userPackage.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * CryptoVault Exchange & Shop API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UserPackage {
  id: string;
  userId: string;
  packageId: string;
  purchasedAt: Date;
  daysCredited: number;
  totalRoiDays: number;
  roiPercent: string;
  principalUsdt: string;
  totalRoiCredited: string;
  isActive: boolean;
  /** @nullable */
  lastCreditedAt?: Date | null;
}